device/requirements.txt
# Requires Python 3.10+
rmscene>=0.6.0
pyyaml>=6.0
